Added access check

+++ src/main/java/com/griddynamics/internship/userservice/model/user/JwtUser.java
@@ -12,7 +12,7 @@ public class JwtUser {
     @NonNull private String accessToken;
     @NonNull private String refreshToken;
     private String type = "Bearer";
-    @NonNull private Long id;
+    @NonNull private Integer id;
     @NonNull private String email;
     @NonNull private String username;
     @NonNull private String role;
